+++ docs/CHANGELOG.md
@@ -3,7 +3,7 @@ title: Changelog
 category: developer
 role: developer
 last_updated: 2025-07-20
-status: draft
+status: complete
 ---
 # Changelog